script set up
