+++ distributions/package.json
@@ -6,9 +6,8 @@
   "license": "MIT",
   "scripts": {
     "start": "CHAIN_ID=1 ts-node ./src/queries/voting-index.ts",
-    "start:rinkeby": "CHAIN_ID=4 ts-node ./src/index.ts",
     "staking": "ts-node ./src/staking.ts",
-    "query": "ts-node ./src/queries/staking-index-direct.ts"
+    "query": "ts-node ./src/queries/staking-index.ts"
   },
   "devDependencies": {
     "ts-node": "^8.10.2",
